Move NavBar component to components directory
